+++ integration/test_runners.py
@@ -1,7 +1,6 @@
 import random
 import time
 
-from edsl.surveys import Survey
 from edsl.questions import QuestionMultipleChoice
 from edsl.scenarios.ScenarioList import ScenarioList
 from edsl.language_models import LanguageModelOpenAIThreeFiveTurbo
@@ -33,30 +32,12 @@ def get_job():
     return q1.by(flip_scenarios).by(m)
 
 
-def test_serial():
-    global serial_time
-    job = get_job()
-    start = time.time()
-    results = job.run(method="serial")
-    end = time.time()
-    serial_time = end - start
-    if verbose:
-        print(f"Serial time: {serial_time}")
-
-
 def test_async():
     global async_time
     job = get_job()
     start = time.time()
-    results = job.run(method="asyncio")
+    results = job.run()
     end = time.time()
     async_time = end - start
     if verbose:
         print(f"Async time: {async_time}")
-
-
-def test_compare():
-    global serial_time, async_time
-    print(f"Serial time: {serial_time}")
-    print(f"Async time: {async_time}")
-    assert async_time < serial_time
